+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 * Added the input `/event/party`. This is the party submitting the
   transaction. You may provide this using the `party` attribute on
   `Minfraud::Components::Event`.
+* Added the input `/payment/method`. This is the payment method associated
+  with the transaction. You may provide this using the `method` attribute
+  on `Minfraud::Components::Payment`.
 
 ## v2.8.0 (2025-05-23)